allow integration tests to automatically rebuild binaries (#18453)
One of the most common pitfalls for me is forgetting to rebuild the
binaries before running the integration tests. There's currently no
automated way to do it, and since the integration tests just pick up the
binaries from $PATH they just use whatever is currently installed on the
system.

This PR introduces a new environment variable that can be set, that
makes the integration tests automatically build the binary into a `bin`
folder in the local repository. This folder is then added to the path,
so `pulumi` and the language hosts from that path are being used instead
of the system version.

Naturally this adds a little bit of time to integration tests, but only
if the envrionment variable is set. Locally to me this seems like a good
tradeoff, since integration tests are most of the time slower than
rebuilding the binary anyway, and most of the time the binaries need to
be rebuilt to run the integration tests.