always disable plugin acquisition for integration tests
Automatic plugin acquisition is disabled for integration tests in CI,
to avoid excessive traffic to GitHub and the associated rate limits.
To avoid confusion when running integration tests locally, set it in
the test main globally for all tests.  This makes the the test
environment more similar to CI, simplifying running tests locally
slightly.