Add crypt bulk methods
We need a pair of methods (supports & encrypt) as we need to know if the crypter supports encryption to change our approach. It also needs to be determinable at runtime rather than via the type system as it might depend on the currently configured system's capabilities.

Implement service manager bulk encryption

Default testSecretsManager to not supporting bulk encryption to maintain existing test expectations.