Implement bulk encryption
Avoid two iterations of the snapshot graph to identify, then fetch, then serialize secrets. Enqueue secrets for encryption which allows the delayed setting of their values.

This also removes the need to manually "insert" secrets into the cache, and also allows us to leverage bulk encryption during the deserialize, if we need it. We should be able to also use the same pattern for the bulk decryption too in the future.

Rename EventualSecretCrypter to bulkSecretEncrypter

Make the interface private to the module as anyone calling from outside wouldn't know to call the finialize method.

Remove ProcessPendingEncryptions in deserialize

We no longer call the enqueue secret encryption during deserialization.

Use explicit bulk operation lifetime

Only enable batching if the caller has explicitly started a batch operation and will therefore complete it.

Serialize secret via pointer

Allow us to set the value of the secret after returning it by avoiding returning a copy of the secret so we can maintain a reference for the bulk process.

Fix deserializeProperty test helper which deserialized back into the source variable which affected the type inference of the json.Unmashal when it was set to a pointer to a struct - causing it to deserialize into the the existing struct rather than defaulting to unpacking into `map[string]any`.