eval: rotate only inputs (#444)
Properties on rotator inputs can now be marked as as RotateOnly, which indicates they are only used during Rotate, not during Open. Rotators must be able to tolerate these properties being passed as unknown during open.

Inline environment references (#443) have been extended to be conditionally evaluated based on the where they are used. During open, evaluation for references that are used by RotateOnly inputs are skipped. This allows a user to use a rotated environment without needing access to managing credentials used by the rotators.

Please note that RotateOnly references will be evaluated during Check of the root environment. This is to ensure that a user who is editing an environment does have permission to access managing credentials.

We should follow this up with something similar to #442 to provide guidance that encourages usage of inline references for RotateOnly imports by emitting diagnostic warnings.

Resolves pulumi/pulumi-service#25240
Resolves pulumi/pulumi-service#25832