Collapse cachingCrypter into cachingSecretsManager
The cachingCrypter has both the encrypter and decrypter fields, but only one of these is set at a time from the manager. This is the only reason we have to pass the encrypter into the deserialize methods - to encrypt when passing into the caching crypter. Therefore, we want to make the caching crypter always have both the encrypter and decrypter fields populated. Or even simpler, just make the cachingSecretsManager implement the crypter inferfaces directly.

Is both Encrypter and Decrypter always available?
- Passphrase, service, cloud and base64 secret managers only have a single crypter object returned from both Encrypter and Decrypter methods.
- The cachingSecretsManager just calls through to the provided Encrypter or Decrypter.
- Only the MockSecretsManager could have only one of the encrypter or decrypter set up.