Fix passing of Name and Type to some provider methods (#18611)
We never saw these in practice because the gRPC layer always fills them
in, but if running engine tests without grpc we'd be missing this data
on the requests.

Noticed this while working on some tests for the new refresh logic and
tried to use `Name` and it was empty.