Add Mise to `pulumi/pulumi` (#18626)
This commit adds [Mise](https://mise.jdx.dev) to the repository as a
means to standardise on the development tooling used for working with
`pulumi/pulumi`. It also stubs out some introductory documentation on it
and cleans up/gathers some of the existing pages we have on contributing
to the repository. As we flesh this out, we should hope to clean out the
cruft from/restructure `CONTRIBUTING.md` to make it more
accessible/point to these pages instead.