eval: fix containsObservableUnknowns being incorrect for non-rotateOnly inputs (#451)
The intent of containsObservableUnknowns is to allow evaluateBuiltinRotate tolerate rotateOnly values being unknown when not rotating. RotateOnly indicates that open will not read those values, so it's okay for them to be unknown.

During rotation, rotate will read all inputs, so all inputs must be known at that time.

Unfortunately the equality isn't correct, because it's returning false for unknown, non-rotateOnly values, which allows for an unknown value to be unexpectedly passed into the provider's Rotate method.

Fixes pulumi/pulumi-service#26101