ast: syntax errors for invalid fn invocation (#449)
Currently when a fn expr is invalid we fall back to parsing it as an object, which results in an object containing a fn:: key, which is otherwise illegal, instead of the expected call. This PR makes these incorrect usages emit syntax errors so that it's easier for the user to identify the mistake ahead of time, instead of after discovering that the function wasn't called.

Case 1 - having an additional key in a fn call:
```
values:
  foo:
    fn::builtin: ...
    anotherKey: "oops"
```

Case 2 - having a fn:: key as a toplevel key in values:
```
values:
  fn::open::provider: ...
```

Resolves #447
---------

Co-authored-by: Pat Gavlin <pat@pulumi.com>