Defer to DiffConfig for if explicit provider resources need replacing (#18509)
The `providerChanged` check in the step generator was really strict for
explicit providers compared to default providers.
For a default provider the engine would call DiffConfig to ask the
provider if resources should be replaced, but for explicit providers the
engine would _always_ trigger a replace.

This removes the check for "default-ness" so that both default and
explicit providers do the same check of calling DiffConfig.

Fixes https://github.com/pulumi/pulumi/issues/18395