Default to UnknownID in providers.NewReference (#18630)
We were a little inconsistent in the tests if we bothered to check if ID
was empty and setting it to UnknownID or if we just passed the ID
directly. There was also a lot of duplicate code doing that empty check.
This moves the empty check to "providers.NewReference" so we can just
always pass ID directly.