[auto] Fix `up` with nested secret in config file
At the end of running an `up` operation from Automation API, the `pulumi
stack history` command is invoked to retrieve an update summary of the
operation.

We want the `stack history --json` command to succeed even if there are
problems decrypting secrets in the config, which can happen when the
`stack history` command doesn't have the encryption salt (because it is
stored in a separate config file). If there is an error decrypting a
config secret, the config's value is replaced with
`ERROR_UNABLE_TO_DECRYPT` rather than having the command fail.

Unfortunately, this does not work when there is a config value with a
nested secret. In that case, an obscure "unexpected end of JSON input"
error is returned and the command fails.

This happens because we attempt to unmarshal the nested value, which is
represented as a JSON string regardless of whether the decrypt succeeded
or failed. In the failure case, the JSON string is an empty string, so
the unmarshal fails with this error.

This change fixes the problem to only unmarshal the nested value if
there were no errors decrypting secrets and the value is not empty.