correctly run versionless plugins in subdirs (#18605)
When we have plugins that are located in a subdir of a repository, we
need to fix up the path to include that subdir, as we install the whole
repository into `~/.pulumi/plugins`, and then use the plugin from a
subdir.

Currently we only do that for "prerelease plugins", which means plugins
whose version is referenced by a commit hash. This is however not
correct, as plugins from Git plugins don't necessarily have a version
number, or they can be referenced by a regular semver version.

Fix up the path unconditionally if we have a subdir instead.

Fixes https://github.com/pulumi/pulumi/issues/18557