[Experimental/components] Skip tests and .files while analyzing Python components (#18689)
This PR changes our Python component analyzer to skip certain files and
directories, in particular `tests` and and file starting with a dot. I
think it's a reasonable default not to expect our users to define any
components in those files.

Resolves https://github.com/pulumi/pulumi/issues/18646. While we could
add a configuration nob for this and make users define the exclusion as
the issue suggest, I believe we want to minimize the number of needed
configuration values to achieve our goal of meeting users where they
are.