[Experimental/Components]: support any type
Support any type.  Unfortunately this is a bit awkward, since we
define it as `export type Input<T> = T | Promise<T> |
OutputInstance<T>;`.  Since `any` can also be a `Promise` or an
`OutputInstance`, the typescript compiler just reduces the type to
`any`, so we can't infer the plain-ness for the schema.