Report all errors back from git clone (#18702)
Currently when we try a git clone for a given url without a reference we
try both master and main. If the clone for master _and_ main both fail
we only report back the error that happened for main.

This changes it so we report both errors. I also added some more log
lines.