improve provider loading based on the name
When creating a new provider, it could be that we either pass in a
provider name, or a local path.  We currently distinguish the two by
checking wheter there is a file separator in the name that's passed
in.

However that misses the convenince of just being able to say
e.g. `pulumi package get-schema .`.  Provider names always follow a
specific pattern, and are not allowed to have `.` in their name.
Instead of just checking for path separators, check whether the name
is a valid provider name, and try to check the local path (where
there's no restrictions on the path) otherwise.