package: also suggest disambiguating for local directories
In the `pulumi package *` command family, we try to load the provider
based on some properties, and if it fails and a corresponding
executable exists, we suggest disambiguating it using "./".

Nowadays we can also have source based plugins, so we should give the
user the same disambiguation option if the name matches a local
directory instead of an executable.  Do that here.