stack/policy rm: require yes flag in non-interactive mode (#18708)
When `--non-interactive` is specified on the command line, pulumi
promises to disable interactive mode for all commands, which means it
should never prompt. This however currently does not work for `policy
rm` and `stack rm`, both of which still prompt even when
`--non-interactive` is specified.

Instead of that, error out when `--non-interactive` is specified, but
`--yes` isn't.

Prompted by
https://github.com/pulumi/pulumi/issues/4717#issuecomment-2680086602.