Error on stack reference import (#18704)
Fixes https://github.com/pulumi/pulumi/issues/18594

We don't expect stack references to be "registered" anymore, and if they
are users actually get a warning telling them to update their SDK. As
such they also shouldn't be imported. Make this case an error to report
to the user rather than a confusing panic.