[sdk/nodejs] Append coverage data
From local testing, when running `make unit_tests` in `sdk/nodejs`, I only get a report for the last run of `yarn run nyc` for the `tests_with_mocks`. The coverage data for the previous command, which contains the majority of unit tests, appears to be missing from the coverage report. Add `--no-clean` so that the coverage data is appended and not cleared out between runs.