[sdk/{go,nodejs,python}] Exclude resource refs from property dependencies for packaged components (#18620)
Part of #13802

The `RegisterResourceRequest` protobuf message includes a
`propertyDependencies` field that is basically a map of property names
to resources that each input property depends on. As an SDK is
serializing a resource's input properties, the serializer keeps track of
dependencies for each property. When an input property value itself is a
resource, the resource is serialized as a "resource reference," (a
special type of object that indicates it is a resource) and the resource
is included as a dependency for the property in `propertyDependencies`.

For custom resources, this is the behavior we want. However, for
packaged components, i.e. multi-language components (MLCs), it is
problematic.

The `propertyDependencies` map is passed along to the component
provider's `Construct` method, and our provider SDKs (Node.js, Go, and
Python) use the `propertyDependencies` to determine whether an instance
of an Output should be created for each property, if the property has
dependencies (and/or if it is a secret).

`propertyDependencies` was initially the only way we had to determine
this in `Construct`, but it didn't work well for properties that had
nested Outputs (inside a nested array or map). In these cases, we don't
want the top-level property to be an Output, we want the leaf nodes that
originally were Outputs to be Outputs. Thus, we created a new type of
special object, "output values," that Outputs could be serialized to so
that the right Output instances could be recreated in `Construct`,
obviating the need for `propertyDependencies`.

A provider can indicate that it supports output values, and if it does,
the engine may send output values to it. However, the provider has no
way of knowing that the engine is sending output values, aside from
looking at the input properties and seeing if there are any Output
values present. When the provider sees an Output value, it ignores
`propertyDependencies`.

This generally works great for Outputs. But there is still some
problematic behavior with resource references, since resource references
are included in `propertyDependencies`. A provider may see the following
input properties (pseudo code):

```
inputs = {
    foo = [resource1, resource2]
}
propertyDependencies = {
    foo = [resource1, resource2]
}
```

- The `foo` property is an array of two resource references, `resource1`
and `resource2`
- The propertyDependencies for `foo` says it depends on `resource1` and
`resource2`

Since there are no output values and `foo` depends on two resources (via
`propertyDependencies`), the provider will create a top-level Output
instance for `foo`: `new Output([resource1, resource2])`. This is not
what we want. We want the deserialized value to be `[resource1,
resource2]`, with no outputs at all.

This change fixes the problem by updating the SDKs to exclude resource
references from `propertyDependencies` for packaged components and calls
to `Call` (resource methods). With a user program using an SDK with
these changes, the `propertyDependencies` sent to the engine (and
therefore to `Construct`) will now look like this:

```
inputs = {
    foo = [resource1, resource2]
}
propertyDependencies = {
    foo = []
}
```

The existing provider SDKs won't create an Output for `foo` because the
`propertyDependencies` for `foo` are empty.

---

Note: This change is a bit different from what I had originally planned.
Originally, I was going to make an engine change that would have stopped
sending `propertyDependencies` to `Construct` (and `Call`) if the
provider said it supported output values. However, in order to make this
work, it required a change to SDKs to indicate that they were sending
output values to the engine (from programs). Otherwise, the engine
wouldn't know that a program was sending output values unless it
detected one in a property value, which might not happen as we saw in
the resource reference example above. With that approach, both an SDK
and CLI update would be required to the get the resource reference fix.
After Fraser's comment last week that we probably shouldn't be including
resource references in `propertyDependencies`, it occurred to me that we
could do that and then only an SDK update would be needed for the fix.
And this doesn't preclude us from phasing out `propertyDependencies` in
the future.

---

This is the Go, Node.js, and Python SDK changes for #13802. Similar
changes for .NET and Java are in progress.

---

The change is broken up into multiple commits for easier reviewing:
1. Add conformance test for components and property deps
2. Generate snapshots for the new conformance test
3. Integration test for Go
4. Node.js SDK
5. Python SDK
6. Go SDK