[sdk/nodejs] Append coverage data (#18716)
From local testing, when running `make unit_tests` in `sdk/nodejs`, and
then subsequently running `yarn run nyc report`, I only see coverage
associated with the `tests_with_mocks`. The coverage data for the
previous `nyc` command, which contains the majority of unit tests,
appears to be missing from the coverage report.

This change adds `--no-clean`, so that the coverage data is appended and
not cleared out between runs.

Similar to the problem we fixed for the Python SDK:
https://github.com/pulumi/pulumi/pull/18475