fix coverage for integration tests
We're currently not calculating coverage from integration tests.  This
seems to stem from the usage of `-coverpkg` with Go, which seems to
not be correct.  It's kinda hard to find good documentation on it, but
it seems like it's not working well with `replace` statements, also
`pkg` should be specified as `./...` instead of the whole package name
to make it work.

Since I couldn't find a way to specify `sdk` properly though, let's
set `coverpkg=all`, which includes all packages in the coverage data.
Hopefully codecov is smart enough to then only include source files in
the repository in its report.