fix sdk/python/Makefile, remove stray file
The dist target in sdk/python/Makefile was wrong, because target local
variables can't just be defined in the target, but have a special
syntax that I was not aware of.  Fix that.

The current version of this accidentally produced a `sdk/python/OBIN`
file, which got checked in, but has no value.  Delete that.