introduce helper function for exiterr, and use it in language tests (#18690)
In Go, `exec.ExitErr` contains the whole Stderr of a command. However
it's `Error()` function unfortunately only returns the exit status,
which in most cases is not enough to debug the problem. In most cases we
want to format these errors including the stderr to make the problems
more debuggable.

Factor out the helper function we already have for the python toolchain,
and make it available to other places as well. Start using it in the Go
conformance tests, where I ran into this today. We should spread the
usage of this to wherever we deal with ExitErrors, but let's start here
for now.