(hopefully) fix flaky TestRunCanceled
There's many possiblilities of what the output is when pulumi up is
canceled, depending on the exact moment the cancelation signal gets
through to the process.  Trying to match against all of them seems
like a bit of a fools errand, so let's only match against the exit
code, which should still give us enough information to check whether
the run was canceled successfully or not.