Try to use latest config when rehydrating default providers from state (#18740)
This changes the engine to try and update default provider config when
rehydrating provider resources from state.

We have to rehydrate a provider state when doing a delete before
replace, when deleting resources at the end of the program, and when
doing refreshes.

With this change when we rehydrate a default provider we'll also look to
see if we can update it's config to whats in the stack config. We first
have to call into diff config to check that the config is compatible
(i.e. not a replace), but as long as its compatible we use that new
configuration rather than the old configuration.

This should give consistent results to when multiple resources are using
a default provider and only one is removed from the program. The other
resources would still trigger the normal config based default provider
registration, and the one being deleted would pick up that fresh
provider instance. Now it gets that config even if nothing actually
registers the default provider.

This affects up, refresh, and destroy.