typescript: import packages from @<namespace>/
Curently when using package add, packages are always namespaced in the
"pulumi" namespace. For python this means they are always named
pulumi_<name>. However it would be better to namespace them by
namespace if that was given, so we can disambiguate them better. This
PR makes that so for TypeScript.