Remove SupportsBulkEncryption from Encrypter
Match the pattern of the decrypter and just use a default implementation if the backend doesn't directly support it.

This was an early optimisation to avoid the double serialization work if we can't do batching. This is no longer required because we capture pointers and push the values back into the serialize result.