use namespace instead of publisher for Go package generation (#18755)
If we don't have a better way to generate the module path for Go in
codegen, we currently use the Publisher field to do so. However the
Publisher field in the schema is free-form text, and doesn't necessarily
fit into a package name. Use the namespace that was set up for this
purpose instead.

Note that this is technically a breaking change, however it's expected
that this is not used very much. This is because most providers would
also set the Repository field, which overrides the Publisher/Namespace.

Part of https://github.com/pulumi/pulumi/issues/18369