Revert "Try to use latest config when rehydrating default providers from state (#18740)" (#18756)
This reverts commit 9fc89cfb166bd102618f917361f3fc4f6942169a.

@VenelinMartinov pointed out this could cause havoc with provider
upgrades.
If a user updates their provider to a new version and updates the config
with it to some setting that is incompatible with the old version than
we can't safely update default provider config for the old versions.

There's no way for us to tell this for sure given just config and state.
So for now we'll have to keep to the old behavior of just using what's
in state.