[sdk/python] Fix `get_organization` returning `None` (#18769)
When versions v3.152.0 and earlier of the `pulumi` python SDK are used
with v3.153.0 of the CLI, the `pulumi.get_organization()` function
returns `None` rather than the expected organization. This is a
regression that was introduced by #18595.

This is what's happening: `pulumi-language-python-exec` ships with the
CLI. In v3.153.0, it tries to create an instance of
`pulumi.runtime.Settings`, passing a new parameter: `root_directory`.
However, older versions of the `pulumi` python SDK do not support this
parameter. An error is raised, which is caught and we fall back to
default values, which doesn't have the `organization`, so it ends up
being `None`.

This change fixes the problem in `pulumi-language-python-exec`. We first
try to create `Settings` with all possible parameters. If it runs into
an error, it falls back to a slower path that inspects the signature of
`Settings.__init__` to determine the supported parameters, and then
tries to create `Settings` again with only the parameters it supports.

Fixes #18768