fix test backend usage in automation test (#18754)
For automation tests we want to use the local backend if
PULUMI_ACCESS_TOKEN is not set. We attempted to do this in this test,
however we didn't pass it to `LocalWorkspace.createStack`, but rather to
the `map` function and failed to realize that.

Pass the test backend to the right function to make this work without
having to set a pulumi access token in the environment.