Implement BatchEncrypter
The BatchEncrypter allows encryption operations to be batched to avoid round-trips for when the underlying encrypter
requires a network call. Encryptions to a secret are enqueued and processed in a batch request either when the queue
is full or when the wrapping transaction is committed. If the cache has an entry for a secret and the plaintext has
not changed, the previous ciphertext is used immediately and not enqueued for the batch operation. Results are
also written to the provided cache.

The CompleteCrypterBatch function must be called to ensure that all enqueued encryption operations are processed.