Add `make test_watch`, a test-watching command for NodeJS (#18750)
I've been doing a lot of automation API changes recently, and it's
really helpful to be able to write a test, then run that test repeatedly
as I make changes. This PR adds `make test_watch` to the `sdk/nodejs`
makefile. There are a few advantages to this over `make test_auto`:

* This runs the Typescript directly using [`tsx`](https://tsx.is/)
instead of compiling the SDK and its tests to JS _before_ running tests.
This means the watch command's startup is much faster.
* It auto-configures a valid organization based on the account with
which you've authenticated. This means you don't get errors when your
account doesn't have the same organizations as the CI does.
* You can run specific tests using the `TEST_GREP` argument, which means
you can run a handful of the tests you care about, rather than waiting
for all of them.

Just as a personal preference, this command has the `--bail` flag.
Normally, if Mocha encounters an error, you can't see what it is _until_
all the tests have been executed. `--bail` forces the suite to exit as
soon as an error is encountered, so you only ever get at most one error.
The disadvantage is that later errors might be more useful in debugging,
but the advantage is that it is significantly faster to figure out what
the problem is when you're dealing with a specific subset of the tests.
Happy to make this a configurable thing later on if anyone has strong
opinions.