correctly marshal pluginspec in GetRequiredPackages (#18786)
When implementing the Git Plugins, we started using the `NewPluginSpec`
function to correctly marshal the source into a plugin name. However we
missed updating GetRequiredPackages with the same functionality. This
means that when e.g. `pulumi install` calls `GetRequiredPackages` to get
the package information, the name will not be what the engine expects it
to be, leading to the package being installed in the wrong location.

Fix that marshalling to fix getting Git plugins via `pulumi install`.

Fixes https://github.com/pulumi/pulumi/issues/18785