Fix plugin "exited prematurely" false positives
When closing a plugin, the plugin's connection state is used to determine whether the plugin had crashed, where a connection state other than `Ready` is presumed crashed. This can lead to false positives, as it's possible for the connection state to change from `Ready` to `Idle` after an idle timeout (e.g. 5 minutes) for plugins that have not crashed.

This change avoids the false positives by making a request to a health check service on the plugin as a way of determining if the plugin is still alive. Most plugins don't and won't implement this service, but that's OK because we treat an unimplemented status as OK. If we get any other error (unavailable, deadline_exceeded, etc.) the plugin is presumed crashed.