nodejs: remove dependency on @opentelemetry/semantic-conventions
We seem to be using this package *only* for importing a couple of
constants and nothing else.  However it causes a huge amount of flakes
in CI for some reason that's not obvious.  Instead of spending a lot
of time to try and figure out why this is flaky, let's just drop this
dependency, as it doesn't provide a lot of value anyway, and we don't
expect these constants to ever really change.