[Experimental/Components]: make provider version optional (#18736)
We don't want to force, in fact we want to discourage, authors from
specifying the version of the plugin, particularly in Git plugins. Make
it optional in the Metadata class, so it no longer needs to be passed
in.

Note that this means `GetPluginInfo` will not return a version in some
cases. We could probably try to grab that version from the Git repo for
each language, but doing that is a bit awkward. Note that in many cases
we don't use the version form `GetPluginInfo` anyway, e.g. in `pulumi
plugin ls` we seem to use the version from the directory in
`~/.pulumi/plugins`.