Print a warning if an output is stringified (#18790)
Python fix for https://github.com/pulumi/pulumi/issues/13518

This adds log warnings to python to warn when an attempt is made to
convert an output to a string.

I'm inclined not to add this behind a flag as cases of this are nearly
always user mistakes that should be fixed.

---------

Co-authored-by: Thomas Gummerer <t.gummerer@gmail.com>